+++ src/lib/components/IconButton.svelte
@@ -2,6 +2,7 @@
   import { createEventDispatcher } from 'svelte'
 
   export let size = '16px'
+  export let smallPadding = false
 
   const dispatch = createEventDispatcher()
 
@@ -21,6 +22,7 @@
 <div
   class="icon-button"
   class:active="{active}"
+  class:sm-p="{smallPadding}"
   style="--joueur-icon-button-size: {size};"
   on:click|stopPropagation
   on:keypress|stopPropagation
@@ -35,10 +37,13 @@
 
 <style>
   .icon-button {
-    --uno: 'cursor-pointer flex items-center hover:bg-primary hover:bg-opacity-8 p-2 rounded-[20px]  relative z-3';
+    --uno: 'cursor-pointer flex items-center hover:bg-primary hover:bg-opacity-8 dark:hover:bg-opacity-20 p-2 rounded-[20px] relative z-3';
     font-size: var(--joueur-icon-button-size);
     user-select: none;
     -webkit-user-select: none;
+  }
+  .sm-p {
+    --uno: 'p-1';
   }
   .active {
     --uno: 'active:bg-opacity-16';