+++ src/lib/icons/WindowMinimize.svelte
@@ -1,0 +1,6 @@
+<svg
+  xmlns="http://www.w3.org/2000/svg"
+  width="1em"
+  height="1em"
+  viewBox="0 0 24 24"><path fill="currentColor" d="M4 20v-2h16v2Z"></path></svg
+>